Refactor(web): Solve circular dependency by relocating functions by context
